♻️ refactor: decompose useFavorites into fetch/optimistic/requestBuilder sub-hooks #11
